+++ R/map_merge.R
@@ -69,6 +69,16 @@ mergeMaps <- function(
   if (!is.list(maps)) stop("Input must be a list of acmap objects", call. = FALSE)
   lapply(maps, check.acmap)
 
+  # Check for duplicate ids before merging
+  duplicated_ags <- unique(unlist(lapply(maps, \(map) agMatchIDs(map)[duplicated(agMatchIDs(map))])))
+  duplicated_srs <- unique(unlist(lapply(maps, \(map) srMatchIDs(map)[duplicated(srMatchIDs(map))])))
+  if (length(duplicated_ags) > 0) {
+    stop(strain_list_error("Cannot merge, at least one of the maps has the following duplicated antigen ids:", duplicated_ags))
+  }
+  if (length(duplicated_srs) > 0) {
+    stop(strain_list_error("Cannot merge, at least one of the maps has the following duplicated serum ids:", duplicated_srs))
+  }
+
   # Set options for any relaxation or optimizations
   optimizer_options <- do.call(RacOptimizer.options, optimizer_options)
   merge_options <- do.call(RacMerge.options, merge_options)
@@ -77,7 +87,7 @@ mergeMaps <- function(
   merge_options$dilution_stepsize <- mean(vapply(maps, dilutionStepsize, numeric(1)))
 
   # Apply the relevant merge method
-  switch(
+  merged_map <- switch(
     method,
     # Table merge
     `table` = {
@@ -133,6 +143,32 @@ mergeMaps <- function(
     # Other merge
     stop(sprintf("Merge type '%s' not recognised", method), call. = FALSE)
   )
+
+  # Merge the groups
+  ag_ids    <- unlist(lapply(maps, agMatchIDs))
+  ag_groups <- unlist(lapply(maps, \(map) {
+    if (is.null(agGroups(map))) rep("", numAntigens(map))
+    else                        as.character(agGroups(map))
+  }))
+
+  sr_ids    <- unlist(lapply(maps, srMatchIDs))
+  sr_groups <- unlist(lapply(maps, \(map) {
+    if (is.null(srGroups(map))) rep("", numSera(map))
+    else                        as.character(srGroups(map))
+  }))
+
+  merged_map_ag_groups <- ag_groups[match(agMatchIDs(merged_map), ag_ids)]
+  merged_map_sr_groups <- sr_groups[match(srMatchIDs(merged_map), sr_ids)]
+
+  if (sum(merged_map_ag_groups != "") > 0) {
+    agGroups(merged_map) <- factor(merged_map_ag_groups, unique(merged_map_ag_groups))
+  }
+  if (sum(merged_map_sr_groups != "") > 0) {
+    srGroups(merged_map) <- factor(merged_map_sr_groups, unique(merged_map_sr_groups))
+  }
+
+  # Return the map
+  merged_map
 
 }